fix: redirect signed-in members away from the join page
Authenticated members have no reason to submit an access request, but
"/join" rendered the request form for them all the same. Split the page
into a server component that redirects signed-in users to /dashboard and
a client JoinForm, matching the landing-page fix and the layout's
auth-cookie short-circuit.
